+++ geomatics/__init__.py
@@ -4,4 +4,4 @@ import geomatics.data
 import geomatics.inspect
 import geomatics.series_1d
 import geomatics.series_2d
-import geomatics.series_ndim
+import geomatics.series_3d
